# apps/api/src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Launchpad API
  version: 1.0.0
  description: API for the Launchpad idea validation platform

servers:
  - url: http://localhost:3001
    description: Development server

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token

  schemas:
    Idea:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
        status:
          type: string
          enum: [draft, published]
        categoryId:
          type: string
        authorId:
          type: string
        voteCount:
          type: integer
        commentCount:
          type: integer
        hasVoted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        color:
          type: string
        ideaCount:
          type: integer

    Comment:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        ideaId:
          type: string
        authorId:
          type: string
        createdAt:
          type: string
          format: date-time

paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK

  /api/ideas:
    get:
      summary: List published ideas
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, most_voted, most_commented]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of ideas
    post:
      summary: Create idea
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, shortDescription, fullDescription, categoryId]
              properties:
                title:
                  type: string
                shortDescription:
                  type: string
                fullDescription:
                  type: string
                categoryId:
                  type: string
                status:
                  type: string
                  enum: [draft, published]
      responses:
        '201':
          description: Created idea

  /api/ideas/trending:
    get:
      summary: Get trending ideas
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, all]
        - name: categoryId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Trending ideas

  /api/ideas/my:
    get:
      summary: Get my ideas
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User ideas

  /api/ideas/{id}:
    get:
      summary: Get idea by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Idea details
        '404':
          description: Not found
    put:
      summary: Update idea
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated idea
    delete:
      summary: Delete idea
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted

  /api/ideas/{id}/vote:
    post:
      summary: Toggle vote
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Vote toggled

  /api/ideas/{id}/comments:
    get:
      summary: Get comments for idea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of comments
    post:
      summary: Add comment
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
      responses:
        '201':
          description: Created comment

  /api/categories:
    get:
      summary: List all categories
      responses:
        '200':
          description: List of categories

  /api/categories/{slug}:
    get:
      summary: Get category by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Category details
        '404':
          description: Not found

  /api/comments/{id}:
    delete:
      summary: Delete comment
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted